+++ src/vrm/gltf/extensions/vrmc_spring_bone.rs
@@ -104,7 +104,7 @@ impl Default for ColliderShape {
 
 impl ColliderShape {
     /// Returns the collision vector from the collider to the target position.
-    pub fn calc_collision(
+    pub fn apply_collision(
         &self,
         next_tail: &mut Vec3,
         collider: &GlobalTransform,
@@ -112,15 +112,14 @@ impl ColliderShape {
         joint_radius: f32,
         bone_length: f32,
     ) {
-        let (scale, _, translation) = collider.to_scale_rotation_translation();
-
-        let max_collider_scale = scale.x.max(scale.y).max(scale.z);
+        let (scale, _, _) = collider.to_scale_rotation_translation();
+        let max_collider_scale = scale.abs().max_element();
         match self {
             Self::Sphere(sphere) => {
-                let translation = translation + Vec3::from(sphere.offset);
+                let translation = collider.transform_point(Vec3::from(sphere.offset));
                 let r = joint_radius + sphere.radius * max_collider_scale;
                 let delta = *next_tail - translation;
-                if delta.norm_squared() <= r * r {
+                if delta.length() <= r * r {
                     let dir = delta.normalize();
                     let pos_from_collider = translation + dir * r;
                     *next_tail = head_global_pos