+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+## Unreleased
+
+### Bug Fixes
+
+- Fixed SpringBone colliders.
+
 ## v0.2.1
 
 [Relsease Notes](https://github.com/not-elm/bevy_vrm1/releases/tag/v0.2.1)